PruebaGlobal/requirements.txt
argcomplete==1.8.2      
asn1crypto==0.24.0     
backcall==0.1.0      
beautifulsoup4    ==       4.5.3      
bleach            ==        3.0.2      
bs4               ==        0.0.1      
certifi            ==       2018.11.29 
cffi               ==       1.11.5     
chardet            ==       3.0.2      
Click              ==       7.0        
cmake              ==       3.13.1     
cryptography        ==      2.4.2      
cycler              ==      0.10.0     
Cython               ==     0.29.2     
decorator           ==      4.3.0      
Django              ==      2.1        
dlib                ==      19.16.0    
docutils            ==      0.14       
docx2txt             ==     0.6        
EbookLib             ==     0.15       
entrypoints           ==    0.2.3      
et-xmlfile            ==    1.0.1      
excel                ==     1.0.0      
face-recognition     ==     1.2.3           
Flask              ==       1.0.2      
futures            ==      3.1.1      
gTTS               ==       2.0.2      
gTTS-token         ==      1.1.3      
idna               ==       2.8        
image              ==       1.5.27     
ipykernel          ==       5.1.0      
ipython            ==       7.2.0      
ipython-genutils   ==       0.2.0      
itsdangerous      ==        1.1.0      
jdcal             ==        1.4        
jedi              ==        0.13.1     
Jinja2             ==       2.10       
jsonschema         ==      2.6.0      
jupyter-client      ==      5.2.3      
jupyter-core        ==      4.4.0      
Kivy-Garden        ==       0.1.4      
kiwisolver         ==       1.0.1      
llvmlite           ==       0.27.0     
lxml              ==        4.3.0      
MarkupSafe        ==        1.1.0      
matplotlib        ==        3.0.2      
mistune           ==        0.8.4      
mkl-fft           ==        1.0.10     
mkl-random        ==        1.0.2      
nbconvert         ==        5.3.1      
nbformat          ==        4.4.0      
notebook          ==        5.7.2      
numba             ==        0.42.0     
numpy            ==         1.15.4     
opencv-python      ==       3.4.4.19   
openpyxl           ==       2.5.12     
pandas             ==       0.23.4     
pandocfilters      ==       1.4.2      
parso              ==       0.3.1      
pexpect           ==        4.6.0      
pickleshare       ==        0.7.5      
Pillow            ==        5.3.0      
pip                 ==      18.1       
prometheus-client    ==     0.4.2      
prompt-toolkit       ==     2.0.7      
protobuf             ==     3.6.1      
ptyprocess          ==      0.6.0      
pycparser          ==       2.19       
pygame             ==       1.9.4      
Pygments           ==       2.2.0      
pymongo             ==      3.7.2      
pyocr               ==      0.5.3      
pyparsing           ==      2.3.0      
PyQt5               ==      5.11.3     
PyQt5-sip           ==      4.19.13    
pytesseract          ==     0.2.6      
python-axolotl        ==    0.1.42     
python-axolotl-curve25519 ==0.4.1.post2
python-crontab       ==     2.3.5      
python-dateutil      ==     2.7.5      
python-pptx         ==      0.6.5      
pyttsx3             ==      2.7        
pytz                 ==     2018.7     
pyzmq               ==      17.1.2     
requests            ==      2.21.0     
scikit-learn         ==     0.20.1     
scipy               ==      1.1.0      
Send2Trash         ==       1.5.0      
setuptools          ==      40.6.2     
six                 ==      1.10.0     
SpeechRecognition     ==    3.6.3      
terminado            ==     0.8.1      
tesseract            ==     0.1.3      
testpath            ==      0.4.2      
tornado             ==      5.1.1      
traitlets           ==      4.3.2      
urllib3             ==      1.24.1     
Wand               ==       0.5.0      
wcwidth            ==       0.1.7      
webencodings       ==       0.5.1      
Werkzeug           ==       0.14.1     
wheel               ==      0.32.3     
wxPython            ==      4.0.3      
xlrd               ==       1.0.0      
XlsxWriter         ==       1.1.2      
xlwt              ==        1.3.0      
yowsup             ==       2.5.7 
